专家抽取小bug修复

+++ szcshl-web/src/main/java/cs/service/expert/ExpertSelConditionServiceImpl.java
@@ -97,26 +97,27 @@ public class ExpertSelConditionServiceImpl implements ExpertSelConditionService 
     @Override
     @Transactional
     public ResultMsg delete(String reviewId,String ids) {
-        ResultMsg resultMsg = null;
         try{
             ExpertReview expertReview = expertReviewRepo.findById(reviewId);
-            int cSize = Validate.isList(expertReview.getExpertSelConditionList())?expertReview.getExpertSelConditionList().size():0;
-            int delLength = StringUtil.getSplit(ids,",").size();
-            //如果是清空抽取条件，则可以重新进行抽取
-            if(cSize > 0 && cSize == delLength){
-                expertReview.setSelectIndex(null);
-                expertReview.setExtractInfo(null);
-                expertReview.setFinishExtract(null);
-                expertReview.setState(null);
-                expertReviewRepo.save(expertReview);
+            if(Validate.isObject(expertReview)){
+                int cSize = Validate.isList(expertReview.getExpertSelConditionList())?expertReview.getExpertSelConditionList().size():0;
+                int delLength = StringUtil.getSplit(ids,",").size();
+                //如果是清空抽取条件，则可以重新进行抽取
+                if(cSize > 0 && cSize == delLength){
+                    expertReview.setSelectIndex(null);
+                    expertReview.setExtractInfo(null);
+                    expertReview.setFinishExtract(null);
+                    expertReview.setState(null);
+                    expertReviewRepo.save(expertReview);
+                }
+                expertSelConditionRepo.deleteById(ExpertSelCondition_.id.getName(), ids);
+                expertSelectedRepo.deleteById(ExpertSelected_.conditionId.getName(),ids);
             }
-            expertSelConditionRepo.deleteById(ExpertSelCondition_.id.getName(), ids);
-            expertSelectedRepo.deleteById(ExpertSelected_.conditionId.getName(),ids);
-            resultMsg = new ResultMsg(true, Constant.MsgCode.OK.getValue(),"删除成功！");
+            return ResultMsg.ok("删除成功！");
         }catch (Exception e){
-            resultMsg = new ResultMsg(false, Constant.MsgCode.ERROR.getValue(),"删除失败！");
+            log.error("删除专家抽取条件异常："+e.getMessage());
+            return ResultMsg.error("删除异常，错误问题已记录，请联系系统管理员处理！");
         }
-        return resultMsg;
     }
 
     /**
